requirements.txt
gradio>=4.0.0
pydantic>=2.0.0
pypdf>=4.0.0
pdf2image>=1.16.0
Pillow>=10.0.0
python-dotenv>=1.0.0
httpx>=0.25.0
pandas>=2.0.0
